requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-braces==1.14.0
django-pwa==1.0.10
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.1
twilio==6.63.0
urllib3==1.26.6
